Merge 3 of 3 (`tuxedo-165-generics`), with the two earlier merges in this branch's history.
Every conflict resolved per ENTRY rather than per side, because none of them was a case of one
branch being right.

**157-native-4x** — four conflicts, all comment-only with identical values on both sides.
Kept this branch's prose where it carried more history (`null_buffer_hoist`'s rows record that
P1 was reverted and is back; theirs did not), and this side's `test_subjects.sh` because it is
a strict superset (`~frame_readers`, `~prealloc_stride`).

**tuxedo-1562-layout-gate** — `parse_type_limit` taken WHOLESALE from loft3-ca.  loft#1630
makes `IntegerSpec::max` an `i64` and a range wholly below zero declarable, which retires the
"upper bound cannot be negative" refusal my side carried; their version already keeps
`limit_refused` in all three places, because I converged on it byte-for-byte this morning.
A clean auto-merge then produced a DUPLICATE `limit_refused: false` initialiser — both sides
had added one in different spots, no conflict raised, and only the compiler caught it.

**tuxedo-165-generics** — the heap register resolved as a per-entry union.  This side read
`OPEN: 4` and theirs `OPEN: 3`, and neither is right for the join: `D-heap-15` closed here,
`D-heap-36` and `D-heap-38` closed there, so only `D-heap-8` and `D-heap-9` are open on both.
**OPEN: 2**, and `rule_tags.py registers` agrees against the entries.

THREE CELLS OF MINE LOST THEIR SUBJECT to loft#1630, and one of them passed VACUOUSLY:
`a_refused_limit_does_not_silence_the_next_declaration` led with `limit(-100, -1) size(1)`,
which now refuses nothing — so `limit_refused` was never set and the cell asserted the second
declaration's error, which it reports either way.  Green, testing nothing.  All three re-based
on the INVERTED range (`limit(10, 3)`), which is the refusal loft#1630 added in its place and
reaches the same recovery.  The `limit_refused` doc comment carried the same dead example and
now says so.

THE RATCHET RE-MEASURED, and the measurement is a finding: `index/optional_ratchet.json` is
derived, so neither side's pin fits the join.  Measured 317 / 1308 and re-pinned.  Chasing the
tests half through a worktree of main, keyed on (file, function, kind) because line numbers
shift: `parse_match_inner` went 4 -> 6 `matches!`, a block loft#1600 duplicated.
⚠ `tuxedo-165-generics` is RED on its own ratchet — pinned 1306, actual 1311 — so those sites
predate this join and are not fixed here; altering loft#1600's arm handling on a merge is not
mine to do blind.  Functions FELL 320 -> 317 against main, which is C127 retiring the
spare-code branches.

CLIPPY, which `find_problems` does not run: 24 warnings, all merge artefacts.  Eleven
`i64::from(…max)` became useless conversions the moment `max` turned `i64` — one of them in
`default_value`, which is C127's own path.  The `signed32()` site keeps its `as` and says why:
it is a `const fn` and `From` is not a const trait (rust-lang#143874), which is exactly why
clippy does not fire there and does on the non-const `i32()` beside it.

AND THE CELL THE JOIN MADE WRITABLE.  `1615`'s ⏳ OWED note is now `negative_default()`:
`limit(-100, -1)` answers **-1** on both backends — the bound nearest zero from BELOW, the
third shape of "the default" and the one a cure reaching for `lo` unconditionally fails.  It
needed C127 for the value and loft#1630 for the declaration, so neither branch could have
written it alone.

+++ src/data.rs
@@ -140,6 +140,9 @@ impl IntegerSpec {
     pub const fn signed32() -> Self {
         IntegerSpec {
             min: i32::MIN + 1,
+            // `as`, not `i64::from`: this is a `const fn` and `From` is not a const trait
+            // yet (rust-lang#143874).  Clippy's cast lint does not fire here for that
+            // reason; it fires on the non-const `i32()` below, which does use `From`.
             max: i32::MAX as i64,
             not_null: false,
             forced_size: None,
@@ -190,7 +193,7 @@ impl IntegerSpec {
     pub fn i32() -> Self {
         IntegerSpec {
             min: i32::MIN + 1,
-            max: i32::MAX as i64,
+            max: i64::from(i32::MAX),
             not_null: false,
             forced_size: NonZeroU8::new(4),
         }
@@ -200,7 +203,7 @@ impl IntegerSpec {
     pub fn u32() -> Self {
         IntegerSpec {
             min: 0,
-            max: i64::from(u32::MAX as i64) - 1,
+            max: i64::from(u32::MAX) - 1,
             not_null: false,
             forced_size: None,
         }
@@ -278,7 +281,7 @@ impl IntegerSpec {
     /// loft#1246 already recorded two of them drifting apart; loft#1254 found the third.
     #[must_use]
     pub fn default_value(&self) -> i64 {
-        let (lo, hi) = (i64::from(self.min), i64::from(self.max));
+        let (lo, hi) = (i64::from(self.min), self.max);
         if lo <= 0 && 0 <= hi {
             0
         } else if lo > 0 {
@@ -294,9 +297,9 @@ impl IntegerSpec {
     #[must_use]
     pub fn usable_max(&self, nullable: bool) -> i64 {
         if self.reserves_narrow_sentinel(nullable) && self.min >= 0 {
-            i64::from(self.max) - 1
+            self.max - 1
         } else {
-            i64::from(self.max)
+            self.max
         }
     }
 
@@ -440,7 +443,7 @@ impl IntegerSpec {
     /// Number of distinct representable values (inclusive range + 1).
     #[must_use]
     pub fn range(&self) -> i64 {
-        i64::from(self.max) - i64::from(self.min) + 1
+        self.max - i64::from(self.min) + 1
     }
 
     /// Can `width` bytes hold every value this range admits? — `@FR-L-Narrow`, read as the
